requirements.txt
fastapi
uvicorn
qdrant-client
fastembed
python-dotenv
asyncpg
openai-agents
SQLAlchemy
pydantic
pydantic-settings
